// Frontend/src/types.ts
type Coordinates = {
  x: number;
  y: number;
};

type paddlesInformations = {
  paddleWidth: number;
  paddleHeight: number;
  playerPaddlePossition: Coordinates;
};

type minMaxSpeed = {
  min: number;
  max: number;
};

type settings = {
  aiSpeed: number;
  ballSpeed: minMaxSpeed;
};

type ballDirection = 1 | -1;

type score = false | "player" | "ai";

export { paddlesInformations, Coordinates, settings, minMaxSpeed, ballDirection, score };